% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformDataFrame.R
\name{asinh_transformation}
\alias{asinh_transformation}
\title{asinh(x / cofactor) transformation on a data frame}
\usage{
asinh_transformation(df, cof)
}
\arguments{
\item{df}{A data frame with elements to be transformed.}

\item{cof}{cofactor}
}
\value{
A data frame with applied transformation.
}
\description{
Perform a arcsinh transformation on every element X of the data frame where, x = asinh(x/cofactor).
}
