% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{cluster}
\alias{cluster}
\title{SCHNEL clustering}
\usage{
cluster(
  file_path,
  dir = F,
  sce = F,
  p_comps = NULL,
  n_scales = 0,
  n_neighbors = 50,
  transformation = "NULL",
  cofactor = 5,
  seeds = -1,
  landmark_treshold = 1.5,
  num_trees = 6,
  num_checks = 1024,
  trans_matrix_prune_treshold = 1.5,
  num_walks = 200,
  num_walks_per_landmark = 200,
  monte_carlo_sampling = TRUE,
  out_of_core_computation = TRUE,
  prop_method = "cluster"
)
}
\arguments{
\item{file_path}{Location of the file(s) to read in.}

\item{dir}{Boolean specifying if the file_path is a directory}

\item{sce}{Boolean specifying whether the input file is a SingleCellExperiment object}

\item{n_scales}{Number of scales in the HSNE hierarchy.}

\item{n_neighbors}{Number of neighbors for kNN graph clustering performed during HSNE hierarchy creation.}

\item{transformation}{Transformation function to be performed on data prior to clustering.}

\item{cofactor}{Optional. In case of an arcsinh transformation on data, the value by which the value of a measurement will be divided by.}

\item{seeds}{Seed parameter for the HSNE dimensionality reduction.}

\item{landmark_treshold}{Minimum value of the connectivity score at which points at a particular scales will be selected as future landmarks.}

\item{num_trees}{Number of trees parameter for the HSNE dimensionality reduction}

\item{num_checks}{Number of checks parameter for the HSNE dimensionality reduction}

\item{num_walks}{Number of random walks performed on transition matrices during the creation of the HSNE heirarchy creation.}

\item{num_walks_per_landmark}{Number of random walks performed from every single landmark during the HSNE hierarchy creation.}

\item{monte_carlo_sampling}{Boolean specifying whether to perform Monte Carlo Method during the HSNE hierarchy creation.}

\item{out_of_core_computation}{Boolean for the out_of_core parameter for the HSNE dimensionality reduction}

\item{prop_method}{Default: cluster; other option: label}

\item{trans_matrix_prune_threshold}{The beta threshold}
}
\description{
Function that runs the SCHNEL algorithm on the specified file(s) and returns a matrix/matrices which rows correspond to datapoints provided
for clustering; columns correspond to scales on which the clustering was performed. The value of a cell represents the cluster to which a given
point belongs to.
}
\examples{
cluster_single_fcs <- cluster("../data/fcs_file.fcs", transformation = "log")
cluster_multiple_files <- cluster("../data/folder_name", dir = T)
}
