% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCSV.R
\name{csv_to_dataframe}
\alias{csv_to_dataframe}
\title{CSV to data frame}
\usage{
csv_to_dataframe(file_path, header = F)
}
\arguments{
\item{file_path}{Location of the file to be read in.}

\item{header}{Boolean specifying whether to keep the header from the csv file.}
}
\value{
Data frame containing values from the .csv file.
}
\description{
Parse csv file into a data frame.
}
