% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFCS.R
\name{fcs_to_dataframe}
\alias{fcs_to_dataframe}
\title{FCS to data frame}
\usage{
fcs_to_dataframe(file_path)
}
\arguments{
\item{file_path}{The Location of the file to read in as data frame.}
}
\value{
Data frame containing data from the .fcs file
}
\description{
Parse data from a .fcs file into a data frame.
}
\details{
The function makes use of a non-Cran\-Bioconductor package ParkerICI/grappolo/
that needs to be installed from GitHub.
}
