% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSNE_reticulate.R
\name{get_num_scales}
\alias{get_num_scales}
\title{Get number of scales in HSNE.}
\usage{
get_num_scales(filepath)
}
\arguments{
\item{filepath}{Path to HSNE file.}
}
\value{
An integer of number of scales.
}
\description{
Get number of scales in HSNE.
}
