% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readH5AD.R
\name{h5ad_to_dataframe}
\alias{h5ad_to_dataframe}
\title{H5AD to data frame}
\usage{
h5ad_to_dataframe(file_path)
}
\arguments{
\item{file_path}{Path to the file to be parsed.}
}
\value{
A dataframe containing data from the h5ad object.
}
\description{
Parse data from a h5ad object into a data frame.
}
\details{
The function makes use of two non-Cran/-Bioconductor packages:
satijalab/seurat-data and mojaveazure/seurat-disk that need to be installed from GitHub.
The packages at the time of the development are under experimental stage if the development.
}
