% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMnist.R
\name{load_mnist_images}
\alias{load_mnist_images}
\title{Load the MNIST images dataset}
\usage{
load_mnist_images(file_path)
}
\arguments{
\item{file_path}{Location of the file with the image samples}
}
\value{
NumericMatrix representation of the dataset
}
\description{
Parse MNIST images from .idx-ubyte format into a NumericMatrix.
}
