% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMnist.R
\name{load_mnist_labels}
\alias{load_mnist_labels}
\title{Load MNIST labels dataset}
\usage{
load_mnist_labels(file_path)
}
\arguments{
\item{file_path}{Location of the file with the image labels.}
}
\value{
A NumericMatrix with one column
}
\description{
Parse MNIST labels from .idx-ubyte format into a NumericMatrix.
}
