% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformDataFrame.R
\name{log_t}
\alias{log_t}
\title{ln(val + 1) function.}
\usage{
log_t(val)
}
\arguments{
\item{val}{a numerical value}
}
\value{
natural logarithm of (val + 1)
}
\description{
Add 1 to the value in case it is 0.
}
