% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformDataFrame.R
\name{log_transformation}
\alias{log_transformation}
\title{log(x + 1) transformation on a data frame}
\usage{
log_transformation(df)
}
\arguments{
\item{df}{A data frame with elements to be transformed.}
}
\value{
A data frame with applied transformation.
}
\description{
Perform a logarithmic transformation on every element X of the data frame where, x = ln(x+1).
}
