% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSNE_reticulate.R
\name{parse_and_cluster}
\alias{parse_and_cluster}
\title{Parses and clusters a HSNE file.}
\usage{
parse_and_cluster(filepath, scale_num, prop_method = "cluster")
}
\arguments{
\item{filepath}{Path to HSNE file.}

\item{scale_num}{Scale number to cluster.}

\item{prop_method}{Leiden algorithm prop method. Either "cluster" or "label"}
}
\value{
A vector of labels.
}
\description{
Parses and clusters a HSNE file.
}
