% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{parse_multiple_to_matrix}
\alias{parse_multiple_to_matrix}
\title{Parse data from files in a specified directory into a NumericalMatrix.}
\usage{
parse_multiple_to_matrix(
  dir_path,
  transformation = F,
  cofactor = 5,
  csv_header = T,
  features_after_pca = 50
)
}
\arguments{
\item{dir_path}{Path to the directory with files to be parsed.}

\item{transformation}{Transformation applied to data prior to clsutering: "log"/"arcsinh"/"FALSE, default = FALSE}

\item{cofactor}{Cofactor used in the arcsinh(x / cofactor) transformation}
}
\description{
Parse data from files in a specified directory into a NumericalMatrix.
}
