% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{parse_to_matrix}
\alias{parse_to_matrix}
\title{Read data into a matrix}
\usage{
parse_to_matrix(
  file_path,
  transformation = F,
  cofactor = 5,
  csv_header = T,
  features_after_pca = 50,
  sce = FALSE
)
}
\arguments{
\item{file_path}{Location of the file to be read in}

\item{transformation}{Transformation applied to data prior to clustering: "log"/"arcsinh"/"FALSE, default = FALSE}

\item{cofactor}{Optional. Cofactor of the arcsinh(x / cofactor) transformation, default = 5}

\item{csv_header}{Boolean specifying if the file contains a header (csv only), default = T}

\item{features_after_pca}{Number of principal components to keep after performing PCA on the data, default = 50}
}
\value{
A NumericMatrix representation of data in the file
}
\description{
Parse the data from the file specified by the file_path into a NumericMatrix.
Takes in .csv, .fcs files as well as objects of type Seurat, h5ad, SingleCellExperiment.
Applies log or asinh transformation with cofactor to data if specified.
}
\examples{
parse_to_matrix("../../data/file_to_read.csv", transformation = "log", csv_header = F)
parse_to_matrix("../../data/file_to_read.fcs", "arcsinh", cofactor = 10)
parse_to_matrix("../../data/file_to_read.rds")
}
