% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run}
\alias{run}
\title{Run the HSNE dimensionality reducion algorithm on a NumericalMatrix.}
\usage{
run(
  nm,
  num_scales,
  seed,
  landmark_threshold,
  num_neighbors,
  num_trees,
  num_checks,
  transition_matrix_prune_thresh,
  num_walks,
  num_walks_per_landmark,
  monte_carlo_sampling,
  out_of_core_computation
)
}
\arguments{
\item{nm}{A numeric matrix on which the algorithm is run.}

\item{num_scales}{Number of scales of the HSNE hierarchy.}
}
\description{
Run the HSNE dimensionality reducion algorithm on a NumericalMatrix.
}
