% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSCE.R
\name{sce_to_dataframe}
\alias{sce_to_dataframe}
\title{Parse a SingleCellExperiment object into a data frame.}
\usage{
sce_to_dataframe(file_path)
}
\arguments{
\item{file_path}{File path to the SingleCellExperiment to be parsed into a data frame.}
}
\value{
A data frame containing data from the SingleCellExperiment object.
}
\description{
Parse a SingleCellExperiment object into a data frame.
}
