% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSeurat.R
\name{seurat_to_dataframe_pca}
\alias{seurat_to_dataframe_pca}
\title{Seurat object to dataframe}
\usage{
seurat_to_dataframe_pca(obj, npcs)
}
\arguments{
\item{obj}{A seurat object from which data will be extracted into a dataframe.}
}
\value{
A dataframe containing the assay data from the Seurat object. Rows correspond to different measurements,
columns to different features
}
\description{
Extract data from the Seurat object into a dataframe.
}
\examples{
## Read in a Seurat object
seu <- readRDS("pbmc3k_final.rds")

> seu
An object of class Seurat
13714 features accross 2638 samples within 1 assay
Active assay: RNA (13714 features, 2000 variable features)
 2 dimensional reductions calculated: pca, umap

## Parse the object into a dataframe
seuM <- seurat_to_dataframe(seu)

> dim(seuM)
[1] 2638 13714

}
